Refactor ducks location
(cherry picked from commit 9bb2e6eb74f23a237c52c7bce2e72a046e5dbdbd)
