+++ src/main/java/jss/bugtorch/mixins/early/minecraft/worldgen/MixinStructureMineshaftPieces$Room.java
@@ -1,14 +1,11 @@
 package jss.bugtorch.mixins.early.minecraft.worldgen;
 
 import java.util.List;
-import java.util.Random;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 
-import jss.bugtorch.core.ducks.IOffsetDuck;
-import net.minecraft.nbt.NBTTagCompound;
-import net.minecraft.world.World;
+import jss.bugtorch.ducks.IOffsetDuck;
 import net.minecraft.world.gen.structure.StructureBoundingBox;
 import net.minecraft.world.gen.structure.StructureComponent;
 import net.minecraft.world.gen.structure.StructureMineshaftPieces;
