+++ src/main/java/jss/bugtorch/mixins/early/minecraft/worldgen/MixinStructureStart.java
@@ -9,7 +9,7 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.LocalCapture;
 
-import jss.bugtorch.core.ducks.IOffsetDuck;
+import jss.bugtorch.ducks.IOffsetDuck;
 import net.minecraft.world.World;
 import net.minecraft.world.gen.structure.StructureComponent;
 import net.minecraft.world.gen.structure.StructureMineshaftPieces;
